test(engine): Move loop validation beside loop derivation
