+++ packages/@n8n/engine/src/graph/graph-validation.error.ts
@@ -1,0 +1,7 @@
+/** Thrown when a graph fails a structural rule and can never execute. */
+export class GraphValidationError extends Error {
+	constructor(message: string) {
+		super(message);
+		this.name = 'GraphValidationError';
+	}
+}
