+++ packages/@n8n/engine/src/graph/index.ts
@@ -11,11 +11,7 @@ export {
 	getPredecessorNodeIds,
 	getSuccessorNodeIds,
 } from './workflow-graph-queries';
-export {
-	GraphValidationError,
-	MAX_SLOT_INDEX,
-	validateExecutableGraph,
-	validateLoops,
-} from './validate-executable-graph';
-export { deriveLoops } from './loops';
+export { GraphValidationError } from './graph-validation.error';
+export { MAX_SLOT_INDEX, validateExecutableGraph } from './validate-executable-graph';
+export { deriveLoops, validateLoops } from './loops';
 export type { WorkflowLoop } from './loops';
